+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - Worker pool abstraction implementation & tests
 - Switch prices to worker pool and remove concurrency for quotes since we support only one quote denom atm.
 - Switch ingest block processing system to rely on worker pool with 2 block processing workers.
+- Wait for cold-start (first block) to be processed before starting the next block to avoid overloading the system.
 
 ## v25.2.0